+++ src/main/java/net/sybyline/scarlet/ScarletUI.java
@@ -784,7 +784,7 @@ public class ScarletUI implements IScarletUI
             FileAnalysis analysis = bundleInfo.analysis;
             VersionedFile versionedFile = bundleInfo.id;
             ModelFile file = bundleInfo.file;
-            FileAnalysisAvatarStats stats = analysis.getAvatarStats();
+            FileAnalysisAvatarStats stats = analysis == null ? null : analysis.getAvatarStats();
             infoStatsAppend(panel, constraints, "Avatar name", ()->avatarDisplayName);
             infoStatsAppend(panel, constraints, "File statistics");
             if (file != null)