+++ src/main/java/net/sybyline/scarlet/ScarletWatchedEntities.java
@@ -17,13 +17,12 @@ import java.util.function.Predicate;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import org.apache.commons.csv.CSVFormat;
-import org.apache.commons.csv.CSVRecord;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.sybyline.scarlet.server.discord.DEnum;
+import net.sybyline.scarlet.util.CsvRecord;
 import net.sybyline.scarlet.util.Func;
 import net.sybyline.scarlet.util.UniqueStrings;
 
@@ -161,11 +160,10 @@ public class ScarletWatchedEntities<E>
         
     }
 
-    @SuppressWarnings("resource")
     public boolean importLegacyCSV(Reader reader, boolean overwrite) throws IOException
     {
         List<WatchedEntity> importedWatchedEntities = new ArrayList<>();
-        for (CSVRecord record : CSVFormat.EXCEL.parse(reader))
+        for (CsvRecord record : CsvRecord.parseDocument(reader))
         {
             WatchedEntity watchedEntity = new WatchedEntity();
             watchedEntity.id = record.get(0);
@@ -183,7 +181,7 @@ public class ScarletWatchedEntities<E>
                 .filter($ -> !$.isEmpty())
                 .map(String::toLowerCase)
                 .forEach(watchedEntity.tags.strings()::add);
-            watchedEntity.critical = Boolean.parseBoolean(record.get(3));
+            watchedEntity.critical = record.getBoolean(3);
             watchedEntity.message = record.get(4);
             importedWatchedEntities.add(watchedEntity);
         }