+++ src/main/java/net/sybyline/scarlet/ext/AvatarSearch_AvtrDB.java
@@ -5,6 +5,7 @@ import java.time.OffsetDateTime;
 
 import javax.annotation.Nullable;
 
+import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
 
 import net.sybyline.scarlet.util.HttpURLInputStream;
@@ -17,6 +18,7 @@ public interface AvatarSearch_AvtrDB
 {
 
     public static int MAX_PAGE_SIZE = 50;
+    public static int MAX_BULK_INGEST = 1000;
     public static final String API_ROOT = "https://api.avtrdb.com/v2";
     public static final String API_ROOT_V3 = "https://api.avtrdb.com/v3";
 
@@ -119,6 +121,12 @@ public interface AvatarSearch_AvtrDB
     {
         public long valid_avatar_ids;
     }
+    public static class IngestResponseV3
+    {
+        public long avatars_enqueued;
+        public long invalid_ids;
+        public String ticket;
+    }
 
     static Long index()
     {
@@ -174,7 +182,7 @@ public interface AvatarSearch_AvtrDB
 
     static RefetchResponse request_refetch(String avatar_id, String token)
     {
-        try (HttpURLInputStream in = HttpURLInputStream.post(API_ROOT+"/avatar/request_refetch", ExtendedUserAgent.init_conn, HttpURLInputStream.writeAsJson(null, null, RefetchRequest.class, new RefetchRequest(avatar_id, token))))
+        try (HttpURLInputStream in = HttpURLInputStream.post(API_ROOT+"/avatar/request_refetch", ExtendedUserAgent.init_conn_json, HttpURLInputStream.writeAsJson(null, null, RefetchRequest.class, new RefetchRequest(avatar_id, token))))
         {
             return in.readAsJson(null, null, RefetchResponse.class);
         }
@@ -187,7 +195,7 @@ public interface AvatarSearch_AvtrDB
 
     static IngestResponse request_ingest(String[] avatar_ids, String attribution)
     {
-        try (HttpURLInputStream in = HttpURLInputStream.post(API_ROOT+"/avatar/ingest", ExtendedUserAgent.init_conn, HttpURLInputStream.writeAsJson(null, null, IngestRequest.class, new IngestRequest(avatar_ids, attribution))))
+        try (HttpURLInputStream in = HttpURLInputStream.post(API_ROOT+"/avatar/ingest", ExtendedUserAgent.init_conn_json, HttpURLInputStream.writeAsJson(null, null, IngestRequest.class, new IngestRequest(avatar_ids, attribution))))
         {
             return in.readAsJson(null, null, IngestResponse.class);
         }
@@ -198,17 +206,17 @@ public interface AvatarSearch_AvtrDB
         }
     }
 
-//    static IngestResponse request_ingest_v3(String[] avatar_ids, String attribution)
-//    {
-//        try (HttpURLInputStream in = HttpURLInputStream.post(API_ROOT_V3+"/avatar/ingest", ExtendedUserAgent.init_conn, HttpURLInputStream.writeAsJson(null, null, IngestRequest.class, new IngestRequest(avatar_ids, attribution))))
-//        {
-//            return in.readAsJson(null, null, IngestResponse.class);
-//        }
-//        catch (IOException ioex)
-//        {
-//            ioex.printStackTrace();
-//            return null;
-//        }
-//    }
+    static IngestResponseV3 request_ingest_v3(String[] avatar_ids, String attribution)
+    {
+        try (HttpURLInputStream in = HttpURLInputStream.post(API_ROOT_V3+"/avatar/ingest", ExtendedUserAgent.init_conn_json, HttpURLInputStream.writeAsJson(null, null, IngestRequest.class, new IngestRequest(avatar_ids, attribution))))
+        {
+            return in.readAsJson(null, null, IngestResponseV3.class);
+        }
+        catch (IOException ioex)
+        {
+            ioex.printStackTrace();
+            return null;
+        }
+    }
 
 }